==> 3 Data Cleaning Legal Status.ipynb <==
import pandas as pd
import numpy as np
import time
from datetime import date
import matplotlib.pyplot as plt
import seaborn as sns
pd.set_option('display.max_columns', None)
import re

# Importing the main dataset

status = pd.read_excel("Raw Data/Legal Status.xlsx", sheet_name = "Tabelle1")
status

Convert columns to snakecase

cols = []
for column in status.columns:
    cols.append(column.lower())
status.columns = cols

status = status.rename(columns={'in-person':'in_person'})

Thats a quick one, ? means the according type of spors betting is legal, everything else means basically not legal.
We map it with 0 for not legal and 1 for legal.

def legal(x):
    if x == '?':
        return 1
    return 0

status['online'] = status['online'].apply(legal)
status['in_person'] = status['in_person'].apply(legal)

status

# Saving

# As csv
status.to_csv('Clean Data/status.csv', index=False)
