==> 4 Data Cleaning Income.ipynb <==
import pandas as pd
import numpy as np
import time
from datetime import date
import matplotlib.pyplot as plt
import seaborn as sns
pd.set_option('display.max_columns', None)
import re

# Importing the main dataset

income = pd.read_excel("Raw Data/Income/Income2018.xlsx", sheet_name = "Data")

income.head(30)

We take the same approach as in notebook 2
We want to extract the following data:

columns = ['year','state','med_income']

data = []

for year in ['2018','2019','2020','2021']:
    income = pd.read_excel("Raw Data/Income/Income"+year+".xlsx", sheet_name = "Data")
    for row in range(2,len(income['Unnamed: 0']),13):
        data.append([
            year,
            income.iloc[row,0],
            income.iloc[row+2,2],
        ])

data

Converting to a dataframe:

income = pd.DataFrame(data,columns=columns)
display(income.head(30))
income.shape

Converting the median income to a number

income['med_income'] = income['med_income'].apply(lambda x: int(x.replace(",","")))
income.head(10)

# Saving

# As csv
income.to_csv('Clean Data/income.csv', index=False)
