==> 5 SQL Save and Read.ipynb <==
# !pip install SQLAlchemy-Utils

import pandas as pd
import numpy as np
import pymysql
from sqlalchemy import create_engine
from sqlalchemy import MetaData
from sqlalchemy_utils import database_exists, create_database
from getpass import getpass
password = getpass()

# Importing all clean files

income = pd.read_csv("Clean Data/income.csv")
wager = pd.read_csv("Clean Data/wager_clean.csv")
status = pd.read_csv("Clean Data/status.csv")
population = pd.read_csv("Clean Data/poverty_population_clean.csv")

# Combining tables
Since population and income have a very similar structure we combine them into a demographics dataframe before uploading into SQL.

demographics = pd.merge(left = population,
                                 right = income,
                                 how = 'inner', 
                                 left_on = ['state','year'], 
                                 right_on= ['state','year'])

demographics

# Adjusting names
### Since / is problematic for usage in pandas, we have to change the name of one column in the wager dataframe

wager.columns

wager = wager.rename(columns={'taxes/jurisdiction_revenue':'taxes'})
wager.columns

### Washington DC
Washington DC is accounted for as Washington DC in the wager dataframe and as District of columbia in the demographics, we have to unify it.

demographics['state'] = demographics['state'].apply(lambda x: 'Washington DC' if x == 'District of Columbia' else x)

demographics['state'].unique()

# Connecting to SQL database
### Creating the engine

connection_string = 'mysql+pymysql://root:'+password+'@localhost/sports_betting'
engine = create_engine(connection_string)

### Creating databse if it doesn't exist yet

if not database_exists(engine.url):
    create_database(engine.url)

# Loading the dataframes into the newly created database

income.to_sql("income", engine, if_exists = "replace")
wager.to_sql("wager", engine, if_exists = "replace")
status.to_sql("status", engine, if_exists = "replace")
population.to_sql("population", engine, if_exists = "replace")
demographics.to_sql("demographics", engine, if_exists = "replace")

# Combining the dataframes and loading from database

While selecting we do a right and a left join with a union statement, because we want to retrieve the data for all states even if there is no gambling allowed.

query = r"""
SELECT 
w.year AS year,
w.month AS month,
w.state AS state,
w.revenue AS revenue,
w.handle AS handle,
w.hold AS hold,
w.taxes AS taxes,
d.population AS population, 
d.poverty AS poverty,
d.unemployed AS unemployed,
d.med_income AS med_income,
s.online AS online,
s.in_person AS in_person
FROM wager w
LEFT JOIN demographics d
ON(w.year = d.year and w.state = d.state)
LEFT JOIN status s
ON(s.state = w.state)

UNION

SELECT 
d2.year AS year,
w2.month AS month,
d2.state AS state,
w2.revenue AS revenue,
w2.handle AS handle,
w2.hold AS hold,
w2.taxes AS taxes,
d2.population AS population, 
d2.poverty AS poverty,
d2.unemployed AS unemployed,
d2.med_income AS med_income,
s2.online AS online,
s2.in_person AS in_person
FROM wager w2
RIGHT JOIN demographics d2
ON(w2.year = d2.year and w2.state = d2.state)
LEFT JOIN status s2
ON(s2.state = d2.state);

            """

df = pd.read_sql_query(query, engine)

df.head(10)

# Handling the new NULL values

Our SQL query created new Null Values. That was to be expected as we did not have the demographic data for 2022.
We will extrapolate the trends in demographics from 2020 to 2021 to have an approximation for the missing demographic data.
Also there will be Null values for the gambling data where sports betting is not legal. Those we can replace with 0.

df.isna().sum()

### Filling 0 revenue for states without gambling

# filling with 0 for states without gambling
for i in ['handle','revenue','hold','taxes']:
    df[i] = df[i].fillna(0)

df.isna().sum()

### The month values
The month values are missing, since we have the population data only on a yearly basis, to not mix it up, we set 0 as a placeholder value.

df['month'] = df['month'].fillna(0)
df.isna().sum()

# Legal status
There were NULL values created for the legal statuses, we have to investigate

display(df[df['online'].isna()].head(10))

# The problem seems to be data for washington DC and Puerto Rico, a short investigation will help us to find out and replace the missing data.
df[df['online'].isna()]['state'].value_counts()

Sports betting is allowed in_person and online in washington DC so we fill in accordingly. For Puerto Rico, it is not a State of the USA and we don't have gambling data for it, so we exclude it from our Analysis.

df.shape

df = df[~ df['state'].isin(['Puerto Rico'])]
df.shape

df['online'] = df['online'].fillna(1)
df['in_person'] = df['in_person'].fillna(1)

df.isna().sum()

# Calculating and inserting the missing population data

# First we find out what is missing
df[df['population'].isna()]['state'].unique()

df[df['population'].isna()]['year'].unique()

As assumed the data for 2022 is missing, we calculate the growth from 2020 to 2021 and apply it for the missing values.

statelist = list(df[df['population'].isna()]['state'].unique())
statelist

# We create a new aggregated dataframe to make the calculation easier
helpframe = df.groupby(['state','year']).agg({'population':np.mean,'poverty':np.mean,'unemployed':np.mean,'med_income':np.mean}).reset_index()

helpframe

helpframe.dtypes

#### Calculating the data

pop22 = []
pov22 = []
une22 = []
inc22 = []

for i in statelist:
    # get value for 2020
    pop20 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2020)]['population'].values[0]
    pov20 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2020)]['poverty'].values[0]
    une20 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2020)]['unemployed'].values[0]
    inc20 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2020)]['med_income'].values[0]
    # get value for 2021
    pop21 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2021)]['population'].values[0]
    pov21 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2021)]['poverty'].values[0]
    une21 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2021)]['unemployed'].values[0]
    inc21 = helpframe[(helpframe['state'] == i) & (helpframe['year'] == 2021)]['med_income'].values[0]
    # inserting the values in the same order as the statelist
    pop22.append(pop21*(pop21/pop20))
    pov22.append(pov21*(pov21/pov20))
    une22.append(une21*(une21/une20))
    inc22.append(inc21*(inc21/inc20))

#### Inserting the data into the main dataframe

df.shape

df.isna().sum()

Since we can't detect the pandas nan with an if clause we first replace all nan with a placeholder 0

df = df.fillna(0)
df.isna().sum()

As wee see all nans are now 0

len(df[df['population']==0])

# Now for every row that contains a 0 we replace the values with our calculated ones
for rownum in range(len(df.index)):
    if df.iloc[rownum]['population'] == 0:
        # We check which position in our lists the state and the corresponding values have
        index = statelist.index(df.iloc[rownum]['state'])
        # We insert the values
        df.at[rownum,'population'] = pop22[index]
        df.at[rownum,'poverty'] = pov22[index]
        df.at[rownum,'unemployed'] = une22[index]
        df.at[rownum,'med_income'] = inc22[index]

As we can see all values are replaced and we have our final dataframe

len(df[df['population']==0])

# Saving the final dataframe to csv and into our database

# As csv
df.to_csv('Clean Data/complete_clean.csv', index=False)

# To database
df.to_sql("complete", engine, if_exists = "replace")

df
