==> 8 Additions.ipynb <==
# This notebook contains additional calculaions

Depending on what I might later on need for my presentation, I will put them in here, to be able to add data without impairing the readability of the previous notebooks

import pandas as pd
import numpy as np
import time
from datetime import date
import matplotlib.pyplot as plt
import seaborn as sns

# Importing the data

df = pd.read_csv("Clean Data/intermediate.csv")

df.head(10)

# Additions

### Reverting to unadjusted numbers
Since tableau nicely adds labels to the numbers it is easier to make the presentation with the original numbers, so I revert the division by milliion

for c in ['revenue', 'handle', 'taxes',
       'population', 'poverty', 'unemployed']:
    df[c] = df[c].apply(lambda x: x*1000000)
df.head(5)

### Seasonality
Since we discovered in the DA, that the Season definitely influences the bettin behaviour we add a column with the season.

def season_calc(month):
    if int(month) in [12,1,2]:
        return 'Winter'
    elif int(month) in [3,4,5]:
        return 'Spring'
    elif int(month) in [6,7,8]:
        return 'Summer'
    elif int(month) in [9,10,11]:
        return 'Autumn'
    else:
        return 'No Data'
    

df['season'] = df['month'].apply(season_calc)

df.tail(5)

### Taxes per Capita
For one graph I need the taxes per capita

df['tax_capita'] = df['taxes']/df['population']
df.head(5)

pd.set_option('display.max_rows', None)
display(df.groupby(['state','year']).agg({'med_income':np.mean}))

### Revenues after taxes
For the presentation

df['rv_aftertax'] = df['revenue'] - df['taxes']
df.tail(5)

df[df['state'] == 'Arkansas']

# Saving the final dataframe

# As csv
df.to_csv('Clean Data/final.csv', index=False)
